refactor: 부스 API 및 쿼리 훅 파일 분리 및 구조 개선
- user/admin API 및 훅 분리
- sharedTypes.ts로 공통 타입 정리
- 디렉토리 구조 명확화
