+++ src/api/booth/user/userBooth.ts
@@ -1,14 +1,15 @@
 import { sendRequest } from "@/api/request";
 import { defaultInstance } from "@/api/instance";
+import type { PubStatus, ReservationInfo } from "../shared/sharedTypes";
 
-// 1) 번호 인증 요청 (/sms/certify)
+// 1. 번호 인증 요청 (/sms/certify)
 export const requestPhoneCert = (phoneNumber: string) => {
   return sendRequest<boolean>(defaultInstance, "POST", "/sms/certify", {
     phoneNumber,
   });
 };
 
-// 2) 번호 인증 검증 (/sms/verify)
+// 2. 인증 코드 검증 (/sms/verify)
 export const verifyPhoneCode = (params: {
   phoneNumber: string;
   certificationNumber: string;
@@ -17,7 +18,7 @@ export const verifyPhoneCode = (params: {
   return sendRequest<boolean>(defaultInstance, "POST", "/sms/verify", params);
 };
 
-// 3) 야간부스 예약 (/reserve/{boothId})
+// 3. 야간부스 예약 (/reserve/{boothId})
 export const reserveBooth = (
   boothId: string,
   data: {
@@ -35,30 +36,18 @@ export const reserveBooth = (
   );
 };
 
-
-// 4) 야간부스 예약 리스트 조회 (/reserve)
-export interface PubStatus {
-  pubsId: number;
-  waitTeam: number;
-  status: "AVAILABLE" | "FULL" | "PREPARING" | "END";
-}
-
+// 4. 전체 부스 상태 조회 (/pubs)
 export const fetchPubsStatus = () => {
   return sendRequest<PubStatus[]>(defaultInstance, "GET", "/pubs");
 };
 
-// 5) 내 예약 정보 조회 (/reserve?number={phoneNumber})
-export interface ReservationInfo {
-  waitTeam: number;
-  reserveStatus: "WAITING" | "DONE" | "CANCELLED";
-}
-
+// 5. 내 예약 정보 조회 (/reserve?number={phoneNumber})
 export const fetchMyReservation = (phoneNumber: string) => {
   const url = `/reserve?number=${phoneNumber}`;
   return sendRequest<ReservationInfo>(defaultInstance, "GET", url);
 };
 
-// 6) 야간부스 예약 취소 (/reserve/cancel)
+// 6. 예약 취소 (/reserve?number={phoneNumber})
 export const cancelReservation = (phoneNumber: string) => {
   const url = `/reserve?number=${phoneNumber}`;
   return sendRequest<boolean>(defaultInstance, "PATCH", url);
